adding value for julia && fix move if dezoom

+++ srcs/colors.c
@@ -6,7 +6,7 @@
 /*   By: afontain <afontain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/21 13:40:07 by afontain          #+#    #+#             */
-/*   Updated: 2023/04/25 14:33:48 by afontain         ###   ########.fr       */
+/*   Updated: 2023/04/25 17:18:19 by afontain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ void	choose_colors(t_data *data, int x, int y, int n)
 	int		b;
 	double	shift;
 
-	shift = data->cs * fabs((100 - 10 * data->count_zoom) / 100);
+	shift = data->cs * fabs((100 - 10 * (double)data->count_zoom) / 100);
 	r = (255.0 * (1.0 - cos(n / 256.0 * 2.0 * M_PI)) / 2.0);
 	g = (255.0 * (1.0 - cos(n / 256.0
 					* shift * 2.0 * M_PI + 2.0 * M_PI / 3.0)) / 2.0);
@@ -47,7 +47,7 @@ void	choose_colors1(t_data *data, int x, int y, int n)
 	int		b;
 	double	shift;
 
-	shift = data->cs * fabs((100 - 10 * data->count_zoom) / 100);
+	shift = data->cs * fabs((100 - 10 * (double)data->count_zoom) / 100);
 	r = (255.0 * (1.0 - cos(n / 256.0
 					* shift * 2.0 * M_PI + 4.0 * M_PI / 3.0)) / 2.0);
 	g = (255.0 * (1.0 - cos(n / 256.0 * 2.0 * M_PI + 1.0 * M_PI / 3.0)) / 2.0);
@@ -62,7 +62,7 @@ void	choose_colors2(t_data *data, int x, int y, int n)
 	int		b;
 	double	shift;
 
-	shift = data->cs * fabs((100 - 10 * data->count_zoom) / 100);
+	shift = data->cs * fabs((100 - 10 * (double)data->count_zoom) / 100);
 	r = (255.0 * (1.0
 				- cos(n / 256.0 * 2.0 * M_PI + 4.0 * M_PI / 3.0)) / 2.0);
 	g = (255.0 * (1.0 - cos(n / 256.0 * 2.0 * M_PI + 1.0 * M_PI / 3.0)) / 2.0);
